Move `Log()` class from main file to framework module
