+++ Main.py
@@ -10,6 +10,7 @@ from discord_slash import SlashCommand, SlashContext
 from discord_slash.utils.manage_commands import create_choice, create_option
 from discord.ext.commands import *
 from discord.ext import tasks
+from framework.logger import Logger
 
 # User Configuration (edit stuff here)
 prefix = "-"  # Default prefix is -, you can replace it with your preferred prefix
@@ -51,52 +52,7 @@ if os.name == "posix":
         open(f"{homedir}/snipe-bot-data/editsnipe.log", 'x', encoding="utf-8")
         open(f"{homedir}/snipe-bot-data/errors.log", 'x', encoding="utf-8")
 
-
-class Log:
-    def __init__(self, os_name: str, directory: str):
-        self.os_name = os_name
-        self.directory = directory
-        start_timestamp = datetime.now().strftime('%H:%M:%S')
-        print(f"[{start_timestamp}] Logging initialized.")
-
-    def snipe(self, text: str):
-        if os.name == "nt":
-            with open(f"{homedir}\\snipe-bot-data\\snipe.log", 'w+', encoding="utf-8") as file:
-                # timestamp = datetime.now().strftime("%H:%M:%S")  Disable internal timestamp logging
-                file.write(f"{text}\n")
-                file.close()
-        elif os.name == "posix":
-            with open(f"{homedir}/snipe-bot-data/snipe.log", 'w+', encoding="utf-8") as file:
-                # timestamp = datetime.now().strftime("%H:%M:%S")  Disable internal timestamp logging
-                file.write(f"{text}\n")
-                file.close()
-
-    def editsnipe(self, text: str):
-        if os.name == "nt":
-            with open(f"{homedir}\\snipe-bot-data\\editsnipe.log", 'w+', encoding="utf-8") as file:
-                # timestamp = datetime.now().strftime("%H:%M:%S")  Disable internal timestamp logging
-                file.write(f"{text}\n")
-                file.close()
-        elif os.name == "posix":
-            with open(f"{homedir}/snipe-bot-data/editsnipe.log", 'w+', encoding="utf-8") as file:
-                # timestamp = datetime.now().strftime("%H:%M:%S")  Disable internal timestamp logging
-                file.write(f"{text}\n")
-                file.close()
-
-    def error(self, text: str):
-        if os.name == "nt":
-            with open(f"{homedir}\\snipe-bot-data\\errors.log", 'w+', encoding="utf-8") as file:
-                # timestamp = datetime.now().strftime("%H:%M:%S")  Disable internal timestamp logging
-                file.write(f"{text}\n")
-                file.close()
-        elif os.name == "posix":
-            with open(f"{homedir}/snipe-bot-data/errors.log", 'w+', encoding="utf-8") as file:
-                # timestamp = datetime.now().strftime("%H:%M:%S")  Disable internal timestamp logging
-                file.write(f"{text}\n")
-                file.close()
-
-
-logger = Log(os.name, homedir)
+logger = Logger(os.name, homedir)
 
 @client.command()
 async def help(ctx):
